# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Qlib Trading Platform API",
    "description": "Quantitative trading with Qlib research engine",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/training/train": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Train Model",
        "description": "Train LightGBM model on historical data.",
        "operationId": "train_model_api_training_train_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/training/models/{model_id}": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Get Model",
        "description": "Get model details.",
        "operationId": "get_model_api_training_models__model_id__get",
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Model Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Model Api Training Models  Model Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/risk/var": {
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "Calculate Var",
        "description": "Calculate Value at Risk at provided confidence.",
        "operationId": "calculate_var_api_risk_var_post",
        "parameters": [
          {
            "name": "portfolio_value",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "title": "Portfolio Value"
            }
          },
          {
            "name": "confidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.95,
              "title": "Confidence"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Calculate Var Api Risk Var Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/risk/sharpe": {
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "Calculate Sharpe",
        "description": "Calculate simplified Sharpe ratio from persisted PnL.",
        "operationId": "calculate_sharpe_api_risk_sharpe_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "risk_free_rate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 0.02,
              "title": "Risk Free Rate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Calculate Sharpe Api Risk Sharpe Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/portfolio/overview": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Portfolio Overview",
        "description": "Get complete portfolio overview.",
        "operationId": "get_portfolio_overview_api_portfolio_overview_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "current_prices",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON string map of symbol->price",
              "title": "Current Prices"
            },
            "description": "JSON string map of symbol->price"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Portfolio Overview Api Portfolio Overview Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/dashboard": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Dashboard",
        "description": "Get full dashboard with all metrics.",
        "operationId": "get_dashboard_api_portfolio_dashboard_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "current_prices",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON string map of symbol->price",
              "title": "Current Prices"
            },
            "description": "JSON string map of symbol->price"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Dashboard Api Portfolio Dashboard Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/performance": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Performance",
        "description": "Get performance metrics.",
        "operationId": "get_performance_api_portfolio_performance_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "current_prices",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "JSON string map of symbol->price",
              "title": "Current Prices"
            },
            "description": "JSON string map of symbol->price"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Performance Api Portfolio Performance Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModelConfigSchema": {
        "properties": {
          "task": {
            "type": "string",
            "title": "Task",
            "description": "classification, multiclass, or regression",
            "default": "classification"
          },
          "target_column": {
            "type": "string",
            "title": "Target Column",
            "default": "label_1d"
          },
          "n_splits": {
            "type": "integer",
            "title": "N Splits",
            "default": 5
          },
          "n_estimators": {
            "type": "integer",
            "title": "N Estimators",
            "default": 100
          },
          "learning_rate": {
            "type": "number",
            "title": "Learning Rate",
            "default": 0.05
          },
          "max_depth": {
            "type": "integer",
            "title": "Max Depth",
            "default": 7
          },
          "num_leaves": {
            "type": "integer",
            "title": "Num Leaves",
            "default": 31
          },
          "early_stopping_rounds": {
            "type": "integer",
            "title": "Early Stopping Rounds",
            "default": 20
          }
        },
        "type": "object",
        "title": "ModelConfigSchema",
        "description": "Model configuration."
      },
...
      "TrainingRequest": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "description": "YYYY-MM-DD"
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "description": "YYYY-MM-DD"
          },
          "config": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelConfigSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "start_date",
          "end_date"
        ],
        "title": "TrainingRequest",
        "description": "Training request."
      },
      "TrainingResult": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "task": {
            "type": "string",
            "title": "Task"
          },
          "n_estimators": {
            "type": "integer",
            "title": "N Estimators"
          },
          "training_date": {
            "type": "string",
            "title": "Training Date"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "feature_count": {
            "type": "integer",
            "title": "Feature Count"
          },
          "top_features": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Top Features"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "task",
          "n_estimators",
          "training_date",
          "model_id",
          "feature_count",
          "top_features"
        ],
        "title": "TrainingResult",
        "description": "Training result."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}